separating pthreads and raw threads and the ability to switch between them. It looks like we won't be able to reasonably use raw threads because there are global values in pthreads that libraries expect to be there. need to fix fonts blending with the compositor

+++ src/cu_std/cu/debugtimer.cpp
@@ -182,7 +182,7 @@ void RecordThread(){
     context->debugtable.thread_count = context->debugtable.thread_count + 1;
     
     printf("Tracking thread index: %d : %d\n",context->debugtable.thread_count - 1,
-           *threadid_ptr);
+           (u32)*threadid_ptr);
     
     TSpinUnlock(&context->locker);
